requirements.txt
cryptography>=39.0.0
Pillow>=9.0.0
tk>=0.1.0